requirements.txt
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.7
django-cors-headers==4.9.0
dnspython==2.8.0
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
msgpack==1.1.2
mysqlclient==2.2.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
zope.interface==8.0.1
